heb_non_payer w correction des erreurs

+++ admins/sup_admin.php
@@ -6,15 +6,25 @@ require_once('../includes/initialize.php');
  
 
 //echo $id."waslaaaaaaaaaat";
-session_start();
-$_SESSION['toast'] = true;
 //redirect_to('index.php');
 
 if(isset($id) && !empty($id)){
 
+    if($id==1)
+    {
+        session_start();
+        $_SESSION['toast_admin'] = true;
+        redirect_to('index.php');
+    }
+    else{
+        session_start();
+        $_SESSION['toast'] = true;
+
         $admins = Admin::delete($id);
         
         redirect_to('index.php');
+    }
+        
         
    
 }else{
